+++ classes/scripts/runner.php
@@ -16,6 +16,8 @@ class runner extends atoum\script\configurable
 {
 	const defaultConfigFile = '.atoum.php';
 	const defaultBootstrapFile = '.bootstrap.atoum.php';
+	const defaultAutoloaderFile = '.autoloader.atoum.php';
+	const defaultComposerAutoloaderFile = 'vendor/autoload.php';
 
 	protected $runner = null;
 	protected $configuratorFactory = null;
@@ -205,6 +207,7 @@ class runner extends atoum\script\configurable
 		# So, this file must be included BEFORE argument parsing which is done in script::run().
 		# Default bootstrap file can be overrided in a default config file included in script\configurable::run() which extends script::run().
 		# So, if a bootstrap file is defined in a default config file, it will be available when arguments on CLI will be parsed
+		$this->setDefaultAutoloaderFiles();
 		$this->setDefaultBootstrapFiles();
 
 		if ($this->autorun() === true && sizeof($this->runner->getDeclaredTestClasses()) > 0)
@@ -477,6 +480,13 @@ class runner extends atoum\script\configurable
 		return $this;
 	}
 
+	public function setAutoloaderFile($autoloaderFile)
+	{
+		$this->runner->setAutoloaderFile($autoloaderFile);
+
+		return $this;
+	}
+
 	public function enableDebugMode()
 	{
 		$this->runner->enableDebugMode();
@@ -571,6 +581,35 @@ class runner extends atoum\script\configurable
 			if ($this->adapter->is_file($defaultBootstrapFile) === true)
 			{
 				$this->setBootstrapFile($defaultBootstrapFile);
+
+				break;
+			}
+		}
+
+		return $this;
+	}
+
+	public function setDefaultAutoloaderFiles($startDirectory = null)
+	{
+		foreach (self::getSubDirectoryPath($startDirectory ?: $this->getDirectory()) as $directory)
+		{
+			$defaultAutoloaderFile = $directory . static::defaultAutoloaderFile;
+
+			if ($this->adapter->is_file($defaultAutoloaderFile) === true)
+			{
+				$this->setAutoloaderFile($defaultAutoloaderFile);
+
+				return $this;
+			}
+		}
+
+		foreach (self::getSubDirectoryPath($startDirectory ?: $this->getDirectory()) as $directory)
+		{
+			$composerAutoloaderFile = $directory . static::defaultComposerAutoloaderFile;
+
+			if ($this->adapter->is_file($composerAutoloaderFile) === true)
+			{
+				$this->setAutoloaderFile($composerAutoloaderFile);
 
 				break;
 			}
@@ -1002,9 +1041,23 @@ class runner extends atoum\script\configurable
 
 						$script->setBootstrapFile($values[0]);
 					},
+					array('-af', '--autoloader-file'),
+					'<file>',
+					$this->locale->_('Include autoloader <file> before executing each test method'),
+					2
+				)
+			->addArgumentHandler(
+					function($script, $argument, $values) {
+						if (sizeof($values) != 1)
+						{
+							throw new exceptions\logic\invalidArgument(sprintf($script->getLocale()->_('Bad usage of %s, do php %s --help for more informations'), $argument, $script->getName()));
+						}
+
+						$script->setBootstrapFile($values[0]);
+					},
 					array('-bf', '--bootstrap-file'),
 					'<file>',
-					$this->locale->_('Include <file> before executing each test method'),
+					$this->locale->_('Include bootstrap <file> before executing each test method'),
 					2
 				)
 			->addArgumentHandler(
@@ -1134,6 +1187,13 @@ class runner extends atoum\script\configurable
 				{
 					$this->verbose($this->locale->_('Using \'%s\' configuration file…', $configFile));
 				}
+			}
+
+			$autoloaderFile = $this->runner->getAutoloaderFile();
+
+			if ($autoloaderFile !== null)
+			{
+				$this->verbose($this->locale->_('Using \'%s\' autoloader file…', $autoloaderFile));
 			}
 
 			$bootstrapFile = $this->runner->getBootstrapFile();