+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # `dev-master`
 
+* [#605](https://github.com/atoum/atoum/pull/605) Automatically include Composer's autoloader if it exists ([@jubianchi], [@agallou])
+* [#605](https://github.com/atoum/atoum/pull/605) Handle `.autoloader.atoum.php` files to define tests autoloader ([@jubianchi])
+* [#605](https://github.com/atoum/atoum/pull/605) Add the `--autoloader-file`/`-af` CLI argument to define which autoloader file to user ([@jubianchi])
+
 # 2.7.0 - 2016-06-20
 
 * [#594](https://github.com/atoum/atoum/pull/594) Add telemtry report to CI builds ([@jubianchi])